backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ReeceJones <reece_jones@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
alembic = "^1.13.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
strawberry-graphql = "^0.236.0"
uvicorn = "^0.30.1"
redis = "^5.0.7"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
httpx = "^0.27.0"
autoflake = "^2.3.1"
cryptography = "^43.0.0"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.2"
pyright = "^1.1.372"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
